requirements.txt
telebot
requests
python-dotenv
psutil
pandas